requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.1
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==3.2.3
django-admin==2.0.1
django-cors-headers==3.1.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.0.2
django-heroku==0.3.1
django-six==1.0.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==19.9.0
idna==2.9
pipenv==2018.11.26
psycopg2==2.8.3
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.2
requests==2.24.0
screen==1.0.1
simplejson==3.16.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.25.9
virtualenv==16.7.5
virtualenv-clone==0.5.3
whitenoise==4.1.3
xlwt==1.3.0
